backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.0
pydantic-settings==2.6.1
python-multipart==0.0.12
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
redis==5.2.0
minio==7.2.10
pillow==11.0.0
python-magic==0.4.27
httpx==0.28.0
aiofiles==24.1.0
prometheus-client==0.21.0
email-validator==2.1.0
